+++ packages/api-types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/api-types
 
+## 2.17.0
+
+### Minor Changes
+
+- 328e7234: extend vessel interface with additional fields
+
+## 2.16.0
+
+### Minor Changes
+
+- d6bc4dc6: include user groups types definition
+
 ## 2.15.0
 
 ### Minor Changes